Extract abstract class CommandTest
